Separate out SLA tuner
